requirements.txt
black==24.4.2
cryptography==43.0.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1